Extracted draw and getBoundingBox methods
Extracted TextItem and ImageItem draw and getBoundingBox methods from their classes and added them to their respective drawers. Created a abstract class called ItemDrawer but i had to add a factory in it and it no longer made sense to have an abstract class. So i divided that abstract class into an interface and a class, those being IItemDrawer interface which has draw and getBoundingBox methods and ItemFactory class which has a static creational method which created the objects according to the type. There are more things i need to improve to the code.