refac: Add strategy pattern to accessors' slide item creation
Strategy pattern was added for when accessors select the kind and content of a slide item. This
was done due to the nature of the implementation inside those accessors, an if else statement was being used and either an imageItem or a textItem was being instantiated according
to a kind variable but this violates the Open and closed principle of S.O.L.I.D. Whenever a new kind of slide item were to be added, changes would have to be made to all accessors
on this specific if else statement, for now there are five accessors, but in case there were more accessors and more slide items this would become a problem.